requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-RESTful==0.3.10
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6
Werkzeug==2.3.0
gunicorn==21.2.0